create project setup
