--- api/api.route.js
@@ -1,7 +1,0 @@
-import { Router } from "express";
-import testRouter from "./test/test.route.js";
-const apiRouter = Router();
-
-apiRouter.use("/test", testRouter);
-
-export default apiRouter;
