--- api/test/test.route.js
@@ -1,7 +1,0 @@
-import { Router } from "express";
-import { getMembers } from "../test/test.service.js";
-const testRouter = Router();
-
-testRouter.get("/", getMembers);
-
-export default testRouter;
