+++ package.json
@@ -21,7 +21,8 @@
   "homepage": "https://github.com/lukagobechia/LMS#readme",
   "dependencies": {
     "cors": "^2.8.5",
-    "express": "^5.1.0"
+    "express": "^5.1.0",
+    "morgan": "^1.10.0"
   },
   "devDependencies": {
     "nodemon": "^3.1.10"
