+++ src/api/apiRouter.js
@@ -1,0 +1,7 @@
+import { Router } from "express";
+// import testRouter from "./test/test.route.js";
+const apiRouter = Router();
+
+// apiRouter.use("/test", testRouter);
+
+export default apiRouter;
